edited user amount type to decimal

+++ tgbot_server/app/model/user_model.py
@@ -1,6 +1,6 @@
 import uuid
 
-from sqlalchemy import Integer, Boolean, String, Float
+from sqlalchemy import Integer, Boolean, String, DECIMAL
 from sqlalchemy.orm import Mapped, mapped_column, relationship
 
 from app.model.base_model import BaseModel
@@ -11,7 +11,7 @@ class User(BaseModel):
 
     tg_id: Mapped[int] = mapped_column(Integer, nullable=False, unique=True)
     sub_uuid: Mapped[str] = mapped_column(String, nullable=False, unique=True, default=lambda: str(uuid.uuid4()))
-    balance: Mapped[float] = mapped_column(Float, default=0.0)
+    balance: Mapped[DECIMAL] = mapped_column(DECIMAL(10, 2), default=0.00)
     is_active: Mapped[bool] = mapped_column(Boolean, default=True)
     is_superuser: Mapped[bool] = mapped_column(Boolean, default=False)
